Changed transform and load logic
I needed to change our schema.sql as well to auto-generate IDs. It would not work otherwise. Given the way I am inserting into the database, there should be no mismatches.

The code before would run fine for a while and then break; hopefully, like this, it is more robust.

The code has been tested locally and on the Docker image.